Update: File structure
updating structure
