+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Poppins } from "next/font/google";
-import "./styles/globals.css";
-import Header from "./header/page";
+import "../styles/globals.css";
+import Header from "../components/header";
 
 const poppins = Poppins({
   subsets: ["latin"],
