+++ src/app/menu/page.tsx
@@ -1,0 +1,12 @@
+import Link from "next/link";
+
+export default function Menu(){
+    return(
+        <div className="flex justify-between text-center max-w-7xl mx-auto py-8 mb-10">
+            {/* Menu part */}
+            <section id="Menu">
+                <p> hello </p>
+            </section>
+        </div>
+    );
+}
