--- src/app/styles/globals.css
@@ -1,36 +1,0 @@
-@import "tailwindcss";
-
-:root {
-  --background: #ffffff;
-  --dark: #212121;
-}
-
-/* @theme inline {
-  --color-background: var(--background);
-  --color-foreground: var(--foreground);
-  --font-sans: var(--font-geist-sans);
-  --font-mono: var(--font-geist-mono);
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    --background: #0a0a0a;
-    --foreground: #ededed;
-  }
-} */
-
-body {
-  font-family: Poppins, Arial, Helvetica, sans-serif;
-}
-
-.btn-retro {
-  border: 2px solid var(--dark);
-  padding: 0.5rem 1rem; /* px-4 py-2 */
-  box-shadow: -4px 4px 0 var(--dark);
-  transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.btn-retro:hover {
-  transform: translate(-4px, 4px); /* move into the shadow space */
-  box-shadow: none; /* shadow disappears */
-}
