+++ src/styles/globals.css
@@ -2,25 +2,39 @@
 
 :root {
   --background: #ffffff;
-  --foreground: #171717;
+  --dark: #212121;
 }
 
-@theme inline {
+/* @theme inline {
   --color-background: var(--background);
   --color-foreground: var(--foreground);
   --font-sans: var(--font-geist-sans);
   --font-mono: var(--font-geist-mono);
 }
 
-@media (prefers-color-scheme: light) {
+@media (prefers-color-scheme: dark) {
   :root {
     --background: #0a0a0a;
     --foreground: #ededed;
   }
-}
+} */
 
 body {
-  background: var(--background);
-  color: var(--foreground);
-  font-family: Arial, Helvetica, sans-serif;
+  font-family: Poppins, Arial, Helvetica, sans-serif;
+}
+
+.btn-retro {
+  border: 2px solid var(--dark);
+  padding: 0.5rem 1rem; /* px-4 py-2 */
+  box-shadow: -4px 4px 0 var(--dark);
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+}
+
+.btn-retro:hover {
+  transform: translate(-4px, 4px); /* move into the shadow space */
+  box-shadow: none; /* shadow disappears */
+}
+
+html {
+  scroll-behavior: smooth;
 }
